added titles and such

+++ public/top_products/embed/by_company_views.php
@@ -3,7 +3,7 @@
 
     echo "<h2>Top 5 Products by Viewcount</h2>";
     $companies = $db->getAllCompanies();
-    $width = (int) 100 / count($companies);
+    $width = 100 / count($companies);
     echo "<table class=\"layout\"><tr>";
     foreach ($companies as $c){
         echo "<td class=\"column\" style=\"border: 1px solid; width: ".$width."%;\">";
